% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adjust_pretest}
\alias{adjust_pretest}
\title{Calculate pretest probability change over time}
\usage{
adjust_pretest(pretest, incubation, asympt)
}
\arguments{
\item{pretest}{Pretest probability (time series)}

\item{incubation}{The distribution of incubation period (days to symptom
onset) of the format
\code{data.frame(x = 0:21, y = dlnorm(c(0:21), 1.63, 0.5))}}

\item{asympt}{The proportion of positive patients who would be expected not
to ever develop symptoms (true asymptomatic patients).}
}
\value{
posttest probability
}
\description{
Calculates the pretest probability over time, assuming the individual does
not develop symptoms, by taking into account the distribution of incubation
periods (defined as the time from exposure to symptom onset).
}
