% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_postest_prob}
\alias{calc_postest_prob}
\title{Calculate posttest probability from pretest probability and test
characteristics}
\usage{
calc_postest_prob(pretest_prob, sens, spec)
}
\arguments{
\item{pretest_prob}{Pretest probability}

\item{sens}{Test sensitivity}

\item{spec}{Test specificity}
}
\value{
posttest probability
}
\description{
Calculate posttest probability from pretest probability and test
characteristics
}
