% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{individual_probability}
\alias{individual_probability}
\title{Calculate a time series of probability for an individual following exposure}
\usage{
individual_probability(test_day, pre0, sens, spec, incubation, asympt)
}
\arguments{
\item{test_day}{Day of PCR test (days since exposure)}

\item{pre0}{Pre-test probability of person on day of exposure}

\item{sens}{A vector of sensitivities by day since exposure}

\item{spec}{The specificity of the PCR test}

\item{incubation}{The distribution of incubation period (days to symptom
onset) of the format
\code{data.frame(x = 0:21, y = dlnorm(c(0:21), 1.63, 0.5))}}

\item{asympt}{The proportion of infected patients expected to remain
asymptomatic throughout the course of infection}
}
\value{
The probability of an event with the specified probability, after n
repetitions
}
\description{
Calculate a time series of probability for an individual following exposure
}
