% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{posttest_series}
\alias{posttest_series}
\title{Calculate post-test probability if testing occurred on each day in a series}
\usage{
posttest_series(pre0, incubation, asympt, sens, spec)
}
\arguments{
\item{pre0}{The pretest probability on day 0 (at exposure)}

\item{incubation}{The distribution of incubation period (days to symptom
onset) of the format
\code{data.frame(x = 0:21, y = dlnorm(c(0:21), 1.63, 0.5))}}

\item{asympt}{The proportion of infected patients expected to remain
asymptomatic throughout the course of infection}

\item{sens}{A vector of sensitivities by day since exposure}

\item{spec}{The test specificity}
}
\value{
A vector of posttest probabilities
}
\description{
Given an initial pretest probability, and assuming symptoms never arise, with
each passing day the pretest probability will be lower, given the person did
not experience symptoms. This returns a vector of posttest probabilities
which takes all of the above into account, assuming a negative test on each
day. Note this is not a time series, and does not reflect if serial testing
were done each day and assumes testing was only done once.
}
