% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prop_remaining}
\alias{prop_remaining}
\title{The remaining individuals who would not be expected to show symptoms yet}
\usage{
prop_remaining(t, incubation, asympt)
}
\arguments{
\item{t}{day}

\item{incubation}{The distribution of incubation period (days to symptom
onset) of the format
\code{data.frame(x = 0:21, y = dlnorm(c(0:21), 1.63, 0.5))}}

\item{asympt}{The proportion of positive patients who would be expected not
to ever develop symptoms (true asymptomatic patients).}
}
\value{
Proportion who would not be expected to show symptoms yet
}
\description{
The remaining individuals who would not be expected to show symptoms yet
}
