% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_data.R
\docType{data}
\name{sens}
\alias{sens}
\title{COVID-19 PCR sensitivity by days since exposure}
\format{
A data frame with 21 rows and 3 variables:
\describe{
\item{point}{point estimate of sensitivity}
\item{lower}{lower 95\% confidence interval of sensitivity}
\item{upper}{upper 95\% confidence interval of sensitivity}
}
}
\source{
\url{https://bmjopen.bmj.com/content/10/8/e039652}
}
\usage{
sens
}
\description{
COVID-19 PCR sensitivity by days since exposure
}
\keyword{datasets}
