% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{individual_probability}
\alias{individual_probability}
\title{Calculate a time series of probability for an individual following exposure}
\usage{
individual_probability(test_day, pre0, sens, spec, asympt, days, mu, sigma)
}
\arguments{
\item{test_day}{Day of PCR test (days since exposure)}

\item{pre0}{Pre-test probability of person on day of exposure}

\item{sens}{A vector of sensitivities by day since exposure}

\item{spec}{The specificity of the PCR test}

\item{asympt}{The proportion of infected patients expected to remain
asymptomatic throughout the course of infection}

\item{days}{The range of days (e.g. 0:14)}

\item{mu}{The mean of a lognormal distribution that approximates the
incubation period for COVID-19. E.g. 1.63 (see reference).}

\item{sigma}{The standard deviation of a lognormal distribution that
approximates the incubation period for COVID-19. E.g. 0.5
(see reference).}
}
\value{
The probability of an event with the specified probability, after n
repetitions
}
\description{
Calculate a time series of probability for an individual following exposure
}
